requirements.txt
Flask==2.3.3
requests==2.31.0
python-dotenv==1.0.0
mysql-connector-python==8.2.0
google-generativeai==0.3.1
Werkzeug==2.3.7
googlemaps==4.10.0
